修改int为长整形

+++ ActFuns.Tools.Excel2Json/JsonExporter.cs
@@ -75,7 +75,7 @@ namespace ActFuns.Tools.Excel2Json
             switch (type)
             {
                 case "int":
-                    value = Convert.ToInt32(data);
+                    value = Convert.ToInt64(data);
                     break;
                 case "float":
                     value = Convert.ToDouble(data);
